requirements.txt
tensorflow==2.14.0
keras==2.14.0
wandb
boto3
s3fs
pandas
numpy==1.24.4
joblib
xgboost 
scikit-learn==1.6.1
fastapi>=0.68.0
uvicorn>=0.15.0
python-dateutil